CORE-1320 file renames; 'apollo.errorHandling' => 'apollo.errorPipeline'
